requirements.txt
opencv-python==4.8.0.74
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
streamlit==1.25.0
tensorflow==2.13.0
tensorflow-cpu==2.13.0